Pass JSONTestSuite input through JSON.decode(data)
We were converting it to a String first to be strict about invalid UTF-8
sequences, since a number of tests expect those to fail, but this
bypasses our UTF-16/UTF-32 detection. Switch back to `JSON.decode(data)`
to handle UTF-16 and UTF-32, and simply override those invalid UTF-8
tests to be considered correct.