Build out support for NSDecimalNumber
The `JSON` type now includes a variant for decimal numbers, allowing the
user to encode them directly without converting them to another type
first, and the normal suite of accessors has been extended where
appropriate to handle decimal numbers. The encoder has also been taught
how to encode decimal numbers.

On platforms that do not support `NSDecimalNumber`, there is a dummy
variant to work around Swift's lack of support for conditional
compilation of switch cases, and the decimal number accessors are
omitted. If the dummy `JSON.decimal` case is used anyway, the encoder
will encode it as `null`.