Add test suite support for the Swift Package Manager
When running in the SPM, our fixtures are read out of the project source
(as we don't have a bundle to read from at runtime).

On Linux, the benchmarks are disabled because they're really slow in the
debug configuration, which is what you get by default with `swift test`.

Note: There are a few test failures on Linux right now. They're all
caused by Swift's broken Decimal implementation.